feat: make header async
